frontend/backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask_sqlalchemy==3.0.2
SQLAlchemy==2.0.21
pyjwt==2.7.0
requests==2.28.2
